+++ responsibleai/responsibleai/_managers/error_analysis_manager.py
@@ -237,6 +237,21 @@ class ErrorAnalysisManager(BaseManager):
         props[Keys.REPORTS] = reports
         return props
 
+    def get_data(self):
+        """Get error analysis data
+
+        :return: A array of ErrorAnalysisConfig.
+        :rtype: List[ErrorAnalysisConfig]
+        """
+        return [
+            self._get_error_analysis(i) for i in self.list()["reports"]]
+
+    def _get_error_analysis(self, report):
+        error_analysis = ErrorAnalysisConfig()
+        error_analysis.maxDepth = report[Keys.MAX_DEPTH]
+        error_analysis.numLeaves = report[Keys.NUM_LEAVES]
+        return error_analysis
+
     @property
     def name(self):
         """Get the name of the error analysis manager.
